+++ CHANGELOG.md
@@ -21,6 +21,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 
 ### Fixed
 
+- Fix error on unknown rules when using `--severity`
 - Remove leading "./" from file paths
 
 ## Version 0.0.6 (2026-02-09)